+++ now/deployment/flow.py
@@ -11,39 +11,16 @@ from kubernetes import config
 from tqdm import tqdm
 from yaspin.spinners import Spinners
 
+from now.apps.base.app import JinaNOWApp
 from now.cloud_manager import is_local_cluster
-from now.constants import JC_SECRET, Modalities
+from now.constants import JC_SECRET
 from now.deployment.deployment import apply_replace, cmd, deploy_wolf
 from now.dialog import UserInput
-from now.finetuning.settings import FinetuneSettings
 from now.log.log import yaspin_extended
 from now.utils import sigmap
 
 cur_dir = pathlib.Path(__file__).parent.resolve()
 _ExecutorConfig = namedtuple('_ExecutorConfig', 'name, uses, uses_with')
-
-
-def get_encoder_config(user_input: UserInput) -> _ExecutorConfig:
-    """
-    Gets the correct Executor running the pre-trained model given the user configuration.
-    :param user_input: Configures user input.
-    :return: Small data-transfer-object with information about the executor
-    """
-    if (
-        user_input.output_modality == Modalities.TEXT_TO_IMAGE
-        or user_input.output_modality == Modalities.TEXT
-    ):
-        return _ExecutorConfig(
-            name='clip',
-            uses=f'jinahub+docker://CLIPEncoder/v0.2.1',
-            uses_with={'pretrained_model_name_or_path': user_input.model_variant},
-        )
-    elif user_input.output_modality == Modalities.MUSIC_TO_MUSIC:
-        return _ExecutorConfig(
-            name='openl3clip',
-            uses=f'jinahub+docker://BiModalMusicTextEncoder',
-            uses_with={},
-        )
 
 
 def batch(data_list, n=1):
@@ -129,52 +106,11 @@ def deploy_k8s(f, ns, num_pods, tmpdir, kubectl_path):
     return gateway_host, gateway_port, gateway_host_internal, gateway_port_internal
 
 
-def get_custom_env_file(
-    user_input: UserInput,
-    finetune_settings: FinetuneSettings,
-    tmpdir,
-):
-    suffix = 'docker' if user_input.deployment_type == 'remote' else 'docker'
-
-    indexer_name = f'jinahub+{suffix}://DocarrayIndexer'
-    encoder_config = get_encoder_config(user_input)
-    linear_head_name = f'jinahub+{suffix}://{finetune_settings.finetuned_model_name}'
-
-    env_file = os.path.join(tmpdir, 'dot.env')
-    with open(env_file, 'w+') as fp:
-        if finetune_settings.bi_modal:
-            pre_trained_embedding_size = (
-                finetune_settings.pre_trained_embedding_size * 2
-            )
-        else:
-            pre_trained_embedding_size = finetune_settings.pre_trained_embedding_size
-        config_string = (
-            f'ENCODER_NAME={encoder_config.uses}\n'
-            f'FINETUNE_LAYER_SIZE={finetune_settings.finetune_layer_size}\n'
-            f'PRE_TRAINED_EMBEDDINGS_SIZE={pre_trained_embedding_size}\n'
-            f'INDEXER_NAME={indexer_name}\n'
-        )
-        if encoder_config.uses_with.get('pretrained_model_name_or_path'):
-            config_string += f'CLIP_MODEL_NAME={encoder_config.uses_with["pretrained_model_name_or_path"]}\n'
-        fp.write(config_string)
-        if finetune_settings.perform_finetuning:
-            fp.write(f'LINEAR_HEAD_NAME={linear_head_name}\n')
-
-    return env_file if env_file else None
-
-
-def get_flow_yaml_name(output_modality: Modalities, finetuning: bool) -> str:
-    options = {
-        Modalities.TEXT_TO_IMAGE: {0: 'flow-clip.yml', 1: 'ft-flow-clip.yml'},
-        Modalities.MUSIC_TO_MUSIC: {1: 'ft-flow-music.yml'},
-        Modalities.TEXT: {0: 'flow-clip.yml'},
-    }
-    return options[output_modality][finetuning]
-
-
 def deploy_flow(
     user_input: UserInput,
-    finetune_settings: FinetuneSettings,
+    app: JinaNOWApp,
+    ns: str,
+    env_file: str,
     index: DocumentArray,
     tmpdir: str,
     kubectl_path: str,
@@ -182,18 +118,8 @@ def deploy_flow(
     from jina import Flow
     from jina.clients import Client
 
-    finetuning = finetune_settings.perform_finetuning
-
-    env_file = get_custom_env_file(user_input, finetune_settings, tmpdir)
-
-    ns = 'nowapi'
-
-    yaml_name = get_flow_yaml_name(user_input.output_modality, finetuning)
-
     if user_input.deployment_type == 'remote':
-        flow = deploy_wolf(
-            path=os.path.join(cur_dir, 'flow', yaml_name), env_file=env_file, name=ns
-        )
+        flow = deploy_wolf(path=app.flow_yaml, env_file=env_file, name=ns)
         host = flow.gateway
         client = Client(host=host)
 
@@ -210,7 +136,7 @@ def deploy_flow(
         from dotenv import load_dotenv
 
         load_dotenv(env_file)
-        f = Flow.load_config(os.path.join(cur_dir, 'flow', yaml_name))
+        f = Flow.load_config(app.flow_yaml)
         (
             gateway_host,
             gateway_port,
@@ -219,7 +145,7 @@ def deploy_flow(
         ) = deploy_k8s(
             f,
             ns,
-            3 + (1 if finetuning else 0),
+            f.num_pods,
             tmpdir,
             kubectl_path=kubectl_path,
         )
